refactor sudoku example to confirm CircuitValue changes are non-breaking
